# quiz/src/main/resources/static/openapi/openapi.yml
openapi: 3.0.0
info:
  title: Quiz Application API
  version: 1.0.0
  description: API for managing quizzes, questions, options, users, and analytics
    with role-based access control.
servers:
  - url: http://localhost:8080
    description: Local development server (HTTP)
security:
  - bearerAuth: []
...
paths:
...
  /quizzes/{quizId}:
    get:
      summary: Get quiz details (excluding sensitive info)
      tags:
      - Quiz Taking
      parameters:
      - name: quizId
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          $ref: '#/components/responses/QuizDetail'
  /quizzes/{quizId}/attempts:
    get:
      security:
      - bearerAuth: []
      summary: "Get current user\u2019s number of attempts on quiz"
      tags:
      - Quiz Taking
      parameters:
      - name: quizId
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Number of attempts
          content:
            application/json:
              schema:
                type: object
                properties:
                  attempts:
                    type: integer
                example:
                  attempts: 1
  /quizzes/{quizId}/start:
    post:
      security:
      - bearerAuth: []
      summary: Start a quiz attempt if under max_attempts limit
      tags:
      - Quiz Taking
      parameters:
      - name: quizId
        in: path
        required: true
        schema:
          type: integer
      responses:
        '201':
          $ref: '#/components/responses/TakeStarted'
        '403':
          description: Max attempts exceeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: "Max attempts exceeded for this quiz"
  /quizzes/{quizId}/submit:
    post:
      security:
      - bearerAuth: []
      summary: Submit answers for an attempt with attempt_number
      tags:
      - Quiz Taking
      parameters:
      - name: quizId
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        $ref: '#/components/requestBodies/SubmitAnswers'
      responses:
        '200':
          $ref: '#/components/responses/SubmitResult'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: "Answers malformed or missing take_id"
  /users/{userId}/results:
    get:
      security:
      - bearerAuth: []
      summary: "View user\u2019s past quiz attempts and scores"
      tags:
      - Quiz Taking
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          $ref: '#/components/responses/UserResults'
  /takes/{takeId}:
    get:
      security:
      - bearerAuth: []
      summary: Get detailed results of a particular attempt
      tags:
      - Quiz Taking
      parameters:
      - name: takeId
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          $ref: '#/components/responses/SubmitResult'
...
  /admin/reports/quiz/{quizId}:
    get:
      security:
      - bearerAuth: []
      summary: Detailed report for specific quiz
      tags:
      - Analytics
      parameters:
      - name: quizId
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          $ref: '#/components/responses/QuizReport'
components:
  schemas:
...
    Quiz:
      type: object
      properties:
        id:
          type: integer
          example: 10
        title:
          type: string
          example: "Java Basics Quiz"
        max_attempts:
          type: integer
          example: 3
        created_by:
          type: integer
          example: 42
      required:
        - id
        - title
        - max_attempts
        - created_by
    Question:
      type: object
      properties:
        id:
          type: integer
          example: 101
        quiz_id:
          type: integer
          example: 10
        text:
          type: string
          example: "Which keyword declares a constant in Java?"
        type:
          type: string
          example: "single-choice"
      required:
        - id
        - quiz_id
        - text
        - type
...
    Take:
      type: object
      properties:
        id:
          type: integer
          example: 5001
        user_id:
          type: integer
          example: 42
        quiz_id:
          type: integer
          example: 10
        attempt_number:
          type: integer
          example: 1
        score:
          type: number
          format: float
          example: 85.5
      required:
        - id
        - user_id
        - quiz_id
        - attempt_number
        - score
    Response:
      type: object
      properties:
        id:
          type: integer
          example: 9001
        take_id:
          type: integer
          example: 5001
        question_id:
          type: integer
          example: 101
        option_id:
          type: integer
          example: 1001
        text:
          type: string
          example: "final"
      required:
        - id
        - take_id
        - question_id
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  requestBodies:
...
    SubmitAnswers:
      description: Submit answers for a take
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              take_id:
                type: integer
              answers:
                type: array
                items:
                  type: object
                  properties:
                    question_id:
                      type: integer
                    option_id:
                      type: integer
            required:
              - take_id
              - answers
          example:
            take_id: 5001
            answers:
              - question_id: 101
                option_id: 1001
              - question_id: 102
                option_id: 1005
  responses:
...
    QuizDetail:
      description: Detailed quiz info
      content:
        application/json:
          schema:
            type: object
            properties:
              quiz:
                $ref: '#/components/schemas/Quiz'
              questions:
                type: array
                items:
                  $ref: '#/components/schemas/Question'
          example:
            quiz:
              id: 10
              title: "Java Basics Quiz"
              max_attempts: 3
              created_by: 42
            questions:
              - id: 101
                quiz_id: 10
                text: "Which keyword declares a constant in Java?"
                type: single-choice
    TakeStarted:
      description: Quiz attempt started
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Take'
          example:
            id: 5001
            user_id: 42
            quiz_id: 10
            attempt_number: 1
            score: 0.0
    SubmitResult:
      description: Submission result with score
      content:
        application/json:
          schema:
            type: object
            properties:
              take:
                $ref: '#/components/schemas/Take'
              responses:
                type: array
                items:
                  $ref: '#/components/schemas/Response'
          example:
            take:
              id: 5001
              user_id: 42
              quiz_id: 10
              attempt_number: 1
              score: 85.5
            responses:
              - id: 9001
                take_id: 5001
                question_id: 101
                option_id: 1001
                text: "final"
    UserResults:
      description: User past attempts and scores
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Take'
          example:
            - id: 5001
              user_id: 42
              quiz_id: 10
              attempt_number: 1
              score: 85.5
            - id: 5002
              user_id: 42
              quiz_id: 11
              attempt_number: 1
              score: 92.0
    QuizReport:
      description: Aggregated quiz report
      content:
        application/json:
          schema:
            type: object
            properties:
              quiz_id:
                type: integer
              average_score:
                type: number
                format: float
              attempts:
                type: integer
            required:
              - quiz_id
              - average_score
              - attempts
          example:
            quiz_id: 10
            average_score: 78.2
            attempts: 124